fix: Refactor types
