+++ mirascope/core/bedrock/_utils/__init__.py
@@ -5,16 +5,12 @@ from ._convert_message_params import convert_message_params
 from ._get_json_output import get_json_output
 from ._handle_stream import handle_stream, handle_stream_async
 from ._setup_call import setup_call
-from ._types import AssistantMessageTypeDef, BedrockMessageParam, UserMessageTypeDef
 
 __all__ = [
-    "AssistantMessageTypeDef",
-    "BedrockMessageParam",
     "calculate_cost",
     "convert_message_params",
     "get_json_output",
     "handle_stream",
     "handle_stream_async",
     "setup_call",
-    "UserMessageTypeDef",
 ]
