+++ mirascope/core/bedrock/_utils/_calculate_cost.py
@@ -5,4 +5,8 @@ def calculate_cost(
     input_tokens: int | float | None, output_tokens: int | float | None, model: str
 ) -> float | None:
     """Calculate the cost of a completion using the Bedrock API."""
+    # NOTE: We are currently investigating a dynamic approach to determine costs
+    # for Bedrock models. This is needed due to the large number of models
+    # available through Bedrock, making it impractical to hardcode all pricing
+    # information in this initial implementation.
     return None
