+++ mirascope/core/bedrock/call_kwargs.py
@@ -3,7 +3,7 @@
 from mypy_boto3_bedrock_runtime.type_defs import ToolTypeDef
 
 from ..base import BaseCallKwargs
-from ._utils._types import BedrockMessageParam
+from ._types import BedrockMessageParam
 from .call_params import BedrockCallParams
