+++ mirascope/core/bedrock/call_response.py
@@ -21,12 +21,8 @@ from types_aiobotocore_bedrock_runtime.type_defs import (
 )
 
 from ..base import BaseCallResponse
+from ._types import AssistantMessageTypeDef, BedrockMessageParam, UserMessageTypeDef
 from ._utils import calculate_cost
-from ._utils._types import (
-    AssistantMessageTypeDef,
-    BedrockMessageParam,
-    UserMessageTypeDef,
-)
 from .call_kwargs import BedrockCallKwargs
 from .call_params import BedrockCallParams
 from .dynamic_config import BedrockDynamicConfig
